+++ src/android/data/sqlite/ConfigurationContract.java
@@ -13,7 +13,6 @@ public final class ConfigurationContract {
         public static final String COLUMN_NAME_NULLABLE = "NULLHACK";
         public static final String COLUMN_NAME_RADIUS = "stationary_radius";
         public static final String COLUMN_NAME_DISTANCE_FILTER = "distance_filter";
-        public static final String COLUMN_NAME_LOCATION_TIMEOUT = "location_timeout";
         public static final String COLUMN_NAME_DESIRED_ACCURACY = "desired_accuracy";
         public static final String COLUMN_NAME_DEBUGGING = "debugging";
         public static final String COLUMN_NAME_NOTIF_TITLE = "notification_title";